requirements.txt
fastapi
uvicorn
langchain_ibm
pinecone-client
streamlit
google-auth-oauthlib
google-api-python-client
python-dotenv
sentence-transformers